+++ Project_Micro/include/mpu6050.h
@@ -30,18 +30,18 @@
 #define MPU6050_ACCEL_ZOUT_L    0x40
 
 typedef struct {
-    int16_t x;
-    int16_t y;
-    int16_t z;
+    int x;
+    int y;
+    int z;
 } gyro_data_t;
 
 typedef struct {
-    int16_t x;
-    int16_t y;
-    int16_t z;
+    int x;
+    int y;
+    int z;
 } accel_data_t;
 
-uint8_t mpu6050_init(void);
+int mpu6050_init(void);
 void mpu6050_read_gyro(gyro_data_t *data);
 void mpu6050_read_accel(accel_data_t *data);
