+++ Project_Micro/include/util.h
@@ -7,7 +7,7 @@
 #include <stdint.h>
 
 /* Function prototypes */
-int16_t abs_int16(int16_t value);
-void int_to_str(int16_t value, char* buffer);
+int abs_int16(int value);
+void int_to_str(int value, char* buffer);
 
 #endif /* UTIL_H */
